Refactor main.js: Move WindowManager and initializeOS to src/system
- Created src/system/window-manager.js for the WindowManager class.
- Created src/system/os-init.js for the initializeOS function and boot logic.
- Cleaned up src/main.js to only include necessary CSS imports and call initializeOS.
- Removed unused imports in src/main.js.
- Renamed WindowManagerSystem to WindowManager for better naming consistency.
